Move AppState to its own file
