+++ crates/re_viewer/src/lib.rs
@@ -4,6 +4,7 @@
 //! including all 2D and 3D visualization code.
 
 mod app;
+mod app_state;
 pub mod blueprint_components;
 pub mod env_vars;
 mod loading;
@@ -15,8 +16,10 @@ mod store_hub;
 mod ui;
 mod viewer_analytics;
 
-use re_log_types::PythonVersion;
-pub(crate) use ui::{memory_panel, selection_panel};
+pub(crate) use {
+    app_state::AppState,
+    ui::{memory_panel, selection_panel},
+};
 
 pub use app::{App, StartupOptions};
 pub use remote_viewer_app::RemoteViewerApp;
@@ -62,7 +65,7 @@ pub fn build_info() -> re_build_info::BuildInfo {
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum AppEnvironment {
     /// Created from the Rerun Python SDK.
-    PythonSdk(PythonVersion),
+    PythonSdk(re_log_types::PythonVersion),
 
     /// Created from the Rerun Rust SDK.
     RustSdk {
